Google phone screen question taken 050416
I was asked to code this in google docs whilst talking through
my process over google hangouts. I did this exact solution except
I had an addition +1 to low when the search value was greated than
the mid which meant I was off by one.

I was asked to do another screener, didn't do well enough to get
through to final round but not poorly enough to be rejected